translation/requirements.txt
torchtext~=0.12.0
torch~=1.11.0
spacy~=3.3.0
wget~=3.2
omegaconf~=2.2.1
marshmallow-dataclass~=8.5.8
hydra-core~=1.2.0
requests~=2.27.1
tqdm~=4.64.0
nltk~=3.7
pandas~=1.4.2
scikit-learn~=1.1.1